fix: align image and surface rect
Previously, the top left of the image was being drawn at the center of
the sprite's collision rectangle. We now align them. Additionally, we no
longer scale the image.

This makes the images really small compared to our screen resolution. We
need to allow us to easily scale both the image and the sprite's surface
in tandem. To do this, we can use self.image.get_rect(), where self.image
is the surface returned by get_image() in the Spritesheet class.